Sorted data integrity for the filenames
Before, if a user would have wanted to have flight data with the same data typed in, the filename would have been the same, which means the same file would have been updated with the new data. Now, if the user makes the request with the same data(same filename), if it at different hours, they are going to be 2 different files with their data.